requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# AI/ML
google-generativeai==0.3.2

# HTTP client for routing
httpx==0.25.2

# Database
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0  # For SQLite async support
asyncpg==0.29.0    # For PostgreSQL async support

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# Utilities
pydantic==2.5.0
python-multipart==0.0.6
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # For test client

# Development
python-dotenv==1.0.0